content/base/test/test_bug465767.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=465767
-->
<head>
  <title>Test for Bug 465767</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=465767">Mozilla Bug 465767</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 465767 **/

function CC() {
  netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");
  window.QueryInterface(Components.interfaces.nsIInterfaceRequestor)
        .getInterface(Components.interfaces.nsIDOMWindowUtils)
        .garbageCollect();
}

function runTest() {
  var node = document.createElement("div");
  var e = null;
  try {
    document.implementation.createDocument("", "", null).adoptNode(node);
    CC();
    document.implementation.createDocument("", "", null).adoptNode(node);
  } catch(ex) {
    e = ex;
  }
  is(e, null, ".adoptNode didn't succeed!");
  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);

</script>
</pre>
</body>
</html>
